requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
psycopg2-binary==2.9.6
redis==4.5.5
requests==2.28.2
